backend/requirements.txt
# Core Flask packages
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.0.0
Flask-APScheduler>=1.12.0

# Essential libraries
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# AI/LLM packages
openai>=1.0.0
anthropic>=0.7.0

# Content processing
markdown>=3.4.0
Pillow>=10.0.0
jieba>=0.42.0

# Database
psycopg2-binary>=2.9.0

# Web server
gunicorn>=21.0.0

# Optional dependencies (for development)
# pandas>=2.0.0
# numpy>=1.24.0
# redis>=5.0.0
# celery>=5.3.0